[SOIN] Extrais le type TraceAudit dans le domaine
